Create inflation_calculator_app.R
I am updating and reworking an old idea I had a while ago. This a calculator app I did about two years ago. I want to update it and make the code a bit more cleaner and more focused in the part of calculating present and past values.
This is just a extremely basic draft. My plan is ending with to scripts, one for the front end and one for the back end. And maybe also a script for css styles.