Present value - value lost
Here I added more of the basic structure of the calculator. Also, the formula for calculating the equivalent money of a value from the past. The money from the past should be more valuable, because of inflation, so here I show how the amount of money one person had in the past is less money today. A lot of work to do yet.
Also added a button and the necessary reactive expressions to prevent the calculator from updating output without hitting the submit.