Big aesthetics update
Well this is a big update, the essence of the alulator is the same. However, I have added a lot of important things to make it usable and also look good. I have to add some boxes for the text that goes below the line plot. But not much more.