Finished app
This is the final version of the app, however server and ui are in the same script. I want to have one script for server and one for ui for the deploying phase. I am moving forward with this change, but i wanted to preserve the 'mono' script that was functional already.